resultsdb reporting: use ELN path for "eln" as well as "ELN"
The fedfind 'release' for ELN composes should really have been
"eln" all along, but it wasn't because of some spectacularly
terrible fedfind code I just fixed. Now we're doing ELN composes
in kojipkgs like we do all the other nightly composes, the
release for it *is* coming up as "eln", and that makes this
break. Let's make it handle both.

Signed-off-by: Adam Williamson <awilliam@redhat.com>