Force QEMUCPU to "Haswell" for ELN jobs
Previously we were doing this in the templates for the flavors
that are specific to ELN composes, but with ELN update testing
that won't work as we're using the same flavor we do on non-ELN
composes. We don't want to bump the CPU for non-ELN tests as it
is good to check that our baseline for non-ELN composes has not
been inadvertently bumped. So, let's bump the CPU only for ELN
tests in the scheduler.

Signed-off-by: Adam Williamson <awilliam@redhat.com>