don't schedule all flavors for updates in cp group with no flavors
I noticed we're scheduling all update test flavors for updates
in the critpath-build group. This is because of a logic error
when the update is only in critpath groups we intend to schedule
no flavors for at all - we wind up on the path intended for
updates that are critpath but don't indicate what critpath
groups they're in. This fixes that so when an update *does* have
the critpath_groups attribute but we associate no flavors with
any of the listed groups, we don't schedule any jobs.

Signed-off-by: Adam Williamson <awilliam@redhat.com>