update scheduling: set BUILDROOT_REPO based on bodhi info
Until now, for update tests, we've configured a buildroot repo
for Rawhide and ELN updates, not for anything else. But we should
really also use a buildroot repo for tests of Branched before
updates-testing activation, when it acts just like Rawhide and
ELN (automatic update creation and push on gating success).

Doing that in the tests is a bit awkward, so let's do it in the
scheduler, and set a BUILDROOT_REPO variable whose value is the
name of the repository directory if a buildroot repo should be
used - if not, the variable is not set.

There's a corresponding commit to the tests which makes it use
this variable for buildroot repo creation, instead of its own
logic.

Signed-off-by: Adam Williamson <awilliam@redhat.com>